Extract tz rebuild into separate file
bugfix #97
